// packages/languages/guu/language/guu-node-type.ts
export const enum GuuASTNodeType {
  ASSIGNMENT = "AssignmentStatement",
  PRINT = "PrintStatement",
  PROCEDURE_DECLARATION = "ProcedureDeclaration",
  BLOCK = "Block",
  CALL = "CallExpression",
  FUNCTION_DECLARATION = "FunctionDeclaration",
  IDENTIFIER = "Identifier",
  NUMERIC_LITERAL = "NumericLiteral",
  SOURCE_FILE = "SourceFile"
}